requirements.txt
pymongo
certifi
python-telegram-bot==13.7
dnspython
requests
python-dotenv